requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
biopython==1.84
Bottleneck==1.3.7
certifi==2024.7.4
cftime==1.6.4
charset-normalizer==3.3.2
chart-studio==1.1.0
click==8.1.7
contourpy==1.2.0
cycler==0.11.0
decorator==4.4.2
docker-pycreds==0.4.0
fasteners==0.17.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
GridDataFormats==1.0.2
gsd==3.3.0
h5py==3.11.0
idna==3.7
igraph==0.11.6
imagecodecs==2023.1.23
imageio==2.33.1
imageio-ffmpeg==0.5.1
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.4
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mda-xdrlib==0.2.0
MDAnalysis==2.7.0
MDAnalysisData==0.9.0
mdtraj==1.10.0
metis==0.2a5
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mmtf-python==1.1.3
moviepy==1.0.3
mpmath==1.3.0
mrcfile==1.5.1
msgpack==1.0.8
multidict==6.0.5
netCDF4==1.6.0
networkx==3.2.1
numba==0.60.0
numexpr==2.8.7
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
platformdirs==4.2.2
plotly==4.9.0
ply==3.11
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyedr==0.8.0
Pympler==1.1
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytng==0.0.0
pytz==2024.1
PyYAML==6.0.1
rdkit==2024.3.3
requests==2.32.3
retrying==1.3.4
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.10.0
setproctitle==1.3.3
setuptools==69.5.1
signatory==1.2.6.1.9.0
sip==6.7.12
six==1.16.0
smmap==5.0.1
sympy==1.13.1
tables==3.9.2
texttable==1.7.0
threadpoolctl==3.5.0
tidynamics==1.1.2
tifffile==2023.4.12
tomli==2.0.1
torch==1.9.0+cu111
torch-cluster==1.5.9
torch_geometric==2.5.3
torch-scatter==2.0.9
torch-sparse==0.6.12
torch-spline-conv==1.2.1
torchaudio==0.9.0
torchdiffeq==0.2.4
torchvision==0.10.0+cu111
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2023.3
unicodedata2==15.1.0
urllib3==2.2.2
wandb==0.17.4
wheel==0.43.0
yarl==1.9.4
zipp==3.17.0
